Update UI and bug fixes
